#3748 Functions cannot be passed directly to Client Components - Fixes in campaignBanner
